Move htmlSpec function to helpers.
